backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.17
pillow==11.0.0
pytest==8.3.4
httpx==0.28.1
